+++ SOPT-Assignment/SOPT-Assignment/Extensions/UICollectionViewCell+Extension.swift
@@ -1,0 +1,14 @@
+//
+//  UICollectionView+Extension.swift
+//  SOPT-Assignment
+//
+//  Created by 이승준 on 11/6/25.
+//
+
+import UIKit
+
+extension UICollectionViewCell {
+    static func identifier() -> String {
+        return String(describing: type(of: self))
+    }
+}
