+++ SOPT-Assignment/SOPT-Assignment/Models/CollectionType.swift
@@ -1,0 +1,12 @@
+//
+//  CollectionType.swift
+//  SOPT-Assignment
+//
+//  Created by 이승준 on 11/7/25.
+//
+
+import Foundation
+
+enum CollectionType {
+    case category, market, banner, rank, categoryCell
+}
