+++ SOPT-Assignment/SOPT-Assignment/Views/BaeminFeedView.swift
@@ -360,107 +360,6 @@ class BaeminFeedView: UIView {
     
 }
 
-enum CollectionType {
-    case category, market, banner, rank, categoryCell
-}
-
-final class BaeminUICollectionView: UICollectionView {
-    var type: CollectionType?
-    
-    override init(frame: CGRect, collectionViewLayout layout: UICollectionViewLayout) {
-        super.init(frame: frame, collectionViewLayout: layout)
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    func setType(_ type: CollectionType) {
-        self.type = type
-    }
-}
-
-final class TapCollectionView: UICollectionView {
-
-    private let underlineView = UIView()
-
-    override init(frame: CGRect, collectionViewLayout layout: UICollectionViewLayout) {
-        super.init(frame: frame, collectionViewLayout: layout)
-        basicSetup()
-        setStyle()
-        setLayout()
-    }
-    
-    private func basicSetup() {
-        self.backgroundColor = .clear
-        self.isScrollEnabled = true
-        self.showsHorizontalScrollIndicator = false
-    }
-
-    private func setStyle() {
-        underlineView.do {
-            $0.backgroundColor = .black
-        }
-    }
-
-    private func setLayout() {
-        self.addSubview(underlineView)
-        
-        underlineView.snp.makeConstraints {
-            $0.height.equalTo(3.5)
-            $0.centerY.equalToSuperview().offset(20)
-            $0.width.equalTo(14)
-            $0.leading.equalToSuperview().inset(15)
-        }
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-struct TapRect {
-    var index: Int
-    var width: CGFloat
-    var xPosition: CGPoint
-}
-
-extension UICollectionView {
-    func fetchCellRectFor(indexPath index: IndexPath,
-        paddingFromLeading: CGFloat, cellHorizontalPadding: CGFloat) -> TapRect {
-        // 1)
-        guard let cellAttributes = self.layoutAttributesForItem(at: index)
-            else { return TapRect(index: 0, width: 0, xPosition: CGPoint()) }
-        // 2)
-        let cellOrigin = cellAttributes.frame.origin
-        // 3)
-        let cellXPosition: CGPoint = CGPoint(
-            x: cellOrigin.x + paddingFromLeading,
-            y: cellOrigin.y)
-        // 4)
-        let cellWidth: CGFloat = cellAttributes.size.width - cellHorizontalPadding
-        
-        return TapRect(index: index.item, width: cellWidth, xPosition: cellXPosition)
-    }
-}
-
-extension TapCollectionView {
-    
-    func moveUnderlineFor(at tappedRect: TapRect, isAnimation: Bool = true) {
-        underlineView.snp.updateConstraints {
-            $0.height.equalTo(3.5)
-            $0.centerY.equalToSuperview().offset(20)
-            $0.width.equalTo(tappedRect.width)
-            $0.leading.equalTo(tappedRect.xPosition.x)
-        }
-        if isAnimation {
-            UIView.animate(withDuration: 0.3) {
-                self.layoutIfNeeded()
-            }
-        }
-    }
-}
-
 extension UIView {
     func setSubviews(_ views: [UIView]) {
         views.forEach({ self.addSubview($0) })
